rename: state 이름 변경 및 컴포넌트 이름 변경
- CheckedFavoriteSection을 CreateContextSection으로 변경
